Added the slug field to the club_cars table, translation, information on the klubni-tacki page
