+++ resources/views/club-cars/partials/car-item.blade.php
@@ -3,10 +3,18 @@
         <div class="img rounded d-flex align-items-end" style="background-image: url('{{ $car->image }}');">
         </div>
         <div class="text text-center">
-            <h2 class="mb-0">{{ $car->title }}</h2>
-            <a href="{{ $car->getUrl() }}" class="btn btn-secondary py-2 mt-2">
-                {{ __t('Детальніше') }}
-            </a>
+            <h2 class="mb-0">{{ $car->t('title') }}</h2>
+            <p class="card-text d-flex">
+                <span class="text-left">{{__t('Модель')}}</span>
+                <span class="ml-auto text-right text-primary">{{ $car->car_model->title . ' ' . $car->getCarGroupTitleAttribute() }}</span>
+            </p>
+            <p class="card-text d-flex">
+                <span class="text-left">{{__t('Кількість дверей')}}</span>
+                <span class="ml-auto text-right text-primary">{{ $car->getCarModelDoorCountAttribute() }}</span>
+            </p>
+        </div>
+        <div class="mt-auto">
+            <a href="{{ $car->getUrl() }}" class="btn btn-primary py-2 w-100 text-center">{{__t('Детальніше')}}</a>
         </div>
     </div>
 </div>
